Extract and rename MarketHoursDatabase.Key class as SecurityDatabaseKey for reuse in SymbolPropertiesDatabase
