+++ Common/Securities/MarketHoursDatabase.cs
@@ -15,7 +15,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 using Newtonsoft.Json;
@@ -35,7 +34,7 @@ namespace QuantConnect.Securities
         private static MarketHoursDatabase _dataFolderMarketHoursDatabase;
         private static readonly object DataFolderMarketHoursDatabaseLock = new object();
 
-        private readonly IReadOnlyDictionary<Key, Entry> _entries;
+        private readonly IReadOnlyDictionary<SecurityDatabaseKey, Entry> _entries;
 
         /// <summary>
         /// Gets an instant of <see cref="MarketHoursDatabase"/> that will always return <see cref="SecurityExchangeHours.AlwaysOpen"/>
@@ -49,7 +48,7 @@ namespace QuantConnect.Securities
         /// <summary>
         /// Gets all the exchange hours held by this provider
         /// </summary>
-        public List<KeyValuePair<Key,Entry>> ExchangeHoursListing
+        public List<KeyValuePair<SecurityDatabaseKey,Entry>> ExchangeHoursListing
         {
             get { return _entries.ToList(); }
         }
@@ -58,7 +57,7 @@ namespace QuantConnect.Securities
         /// Initializes a new instance of the <see cref="MarketHoursDatabase"/> class
         /// </summary>
         /// <param name="exchangeHours">The full listing of exchange hours by key</param>
-        public MarketHoursDatabase(IReadOnlyDictionary<Key, Entry> exchangeHours)
+        public MarketHoursDatabase(IReadOnlyDictionary<SecurityDatabaseKey, Entry> exchangeHours)
         {
             _entries = exchangeHours.ToDictionary();
         }
@@ -155,11 +154,11 @@ namespace QuantConnect.Securities
         public virtual Entry GetEntry(string market, string symbol, SecurityType securityType, DateTimeZone overrideTimeZone = null)
         {
             Entry entry;
-            var key = new Key(market, symbol, securityType);
+            var key = new SecurityDatabaseKey(market, symbol, securityType);
             if (!_entries.TryGetValue(key, out entry))
             {
                 // now check with null symbol key
-                if (!_entries.TryGetValue(new Key(market, null, securityType), out entry))
+                if (!_entries.TryGetValue(new SecurityDatabaseKey(market, null, securityType), out entry))
                 {
                     if (securityType == SecurityType.Base)
                     {
@@ -211,111 +210,6 @@ namespace QuantConnect.Securities
             {
                 DataTimeZone = dataTimeZone;
                 ExchangeHours = exchangeHours;
-            }
-        }
-
-        /// <summary>
-        /// Represents the key to a single entry in the <see cref="MarketHoursDatabase"/>
-        /// </summary>
-        public class Key : IEquatable<Key>
-        {
-            private const string Wildcard = "[*]";
-
-            /// <summary>
-            /// The market. If null, ignore market filtering
-            /// </summary>
-            public readonly string Market;
-            /// <summary>
-            /// The symbol. If null, ignore symbol filtering
-            /// </summary>
-            public readonly string Symbol;
-            /// <summary>
-            /// The security type
-            /// </summary>
-            public readonly SecurityType SecurityType;
-            /// <summary>
-            /// Initializes a new instance of the <see cref="Key"/> class
-            /// </summary>
-            /// <param name="market">The market</param>
-            /// <param name="symbol">The symbol. specify null to apply to all symbols in market/security type</param>
-            /// <param name="securityType">The security type</param>
-            public Key(string market, string symbol, SecurityType securityType)
-            {
-                Market = market;
-                SecurityType = securityType;
-                Symbol = symbol;
-            }
-
-            /// <summary>
-            /// Parses the specified string as a <see cref="Key"/>
-            /// </summary>
-            /// <param name="key">The string representation of the key</param>
-            /// <returns>A new <see cref="Key"/> instance</returns>
-            public static Key Parse(string key)
-            {
-                var parts = key.Split('-');
-                if (parts.Length != 3)
-                {
-                    throw new ArgumentException("The specified key was not in the expected format: " + key);
-                }
-                SecurityType type;
-                if (!Enum.TryParse(parts[0], out type))
-                {
-                    throw new ArgumentException("Unable to parse '" + parts[2] + "' as a SecurityType.");
-                }
-
-                var market = parts[1];
-                if (market == Wildcard) market = null;
-
-                var symbol = parts[2];
-                if (symbol == Wildcard) symbol = null;
-
-                return new Key(market, symbol, type);
-            }
-
-            #region Equality members
-
-            public bool Equals(Key other)
-            {
-                if (ReferenceEquals(null, other)) return false;
-                if (ReferenceEquals(this, other)) return true;
-                return string.Equals(Market, other.Market) && Equals(Symbol, other.Symbol) && SecurityType == other.SecurityType;
-            }
-
-            public override bool Equals(object obj)
-            {
-                if (ReferenceEquals(null, obj)) return false;
-                if (ReferenceEquals(this, obj)) return true;
-                if (obj.GetType() != this.GetType()) return false;
-                return Equals((Key)obj);
-            }
-
-            public override int GetHashCode()
-            {
-                unchecked
-                {
-                    var hashCode = (Market != null ? Market.GetHashCode() : 0);
-                    hashCode = (hashCode * 397) ^ (Symbol != null ? Symbol.GetHashCode() : 0);
-                    hashCode = (hashCode * 397) ^ (int)SecurityType;
-                    return hashCode;
-                }
-            }
-
-            public static bool operator ==(Key left, Key right)
-            {
-                return Equals(left, right);
-            }
-
-            public static bool operator !=(Key left, Key right)
-            {
-                return !Equals(left, right);
-            }
-
-            #endregion
-
-            public override string ToString()
-            {
-                return string.Format("{0}-{1}-{2}", SecurityType, Market ?? Wildcard, Symbol ?? Wildcard);
             }
         }
